Chore/upgrade dotnet10

+++ backend/src/HorseRacing.Application/Features/BettingEngine/DTOs/PlaceBetRequest.cs
@@ -2,7 +2,7 @@ namespace HorseRacing.Application.Features.BettingEngine.DTOs;
 
 public class PlaceBetRequest
 {
-    public int RaceId { get; set; }
+    public long RaceId { get; set; }
     public int HorseId { get; set; }
     public decimal Amount { get; set; }
 }
