+++ backend/src/HorseRacing.Application/Features/BettingEngine/Interfaces/IBetRepository.cs
@@ -1,6 +1,7 @@
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using HorseRacing.Domain.Entities;
+using HorseRacing.Domain.Entities.Tournaments;
 
 namespace HorseRacing.Application.Features.BettingEngine.Interfaces;
 
@@ -9,22 +10,22 @@ public interface IBetRepository
     // Bet operations
     Task<Bet?> GetByIdAsync(int id);
     Task<IEnumerable<Bet>> GetByUserIdAsync(int userId);
-    Task<IEnumerable<Bet>> GetByRaceIdAsync(int raceId);
+    Task<IEnumerable<Bet>> GetByRaceIdAsync(long raceId);
     Task AddAsync(Bet bet);
     Task SaveChangesAsync();
 
     // Race queries
-    Task<Race?> GetRaceByIdAsync(int raceId);
-    Task<bool> IsHorseInRaceAsync(int raceId, int horseId);
-    Task<RaceResult?> GetRaceResultAsync(int raceId);
-    Task<RaceEntry?> GetRaceEntryAsync(int raceId, int horseId);
+    Task<Race?> GetRaceByIdAsync(long raceId);
+    Task<bool> IsHorseInRaceAsync(long raceId, int horseId);
+    Task<RaceResult?> GetRaceResultAsync(long raceId);
+    Task<RaceEntry?> GetRaceEntryAsync(long raceId, int horseId);
     Task<Horse?> GetHorseByIdOrNameAsync(string identifier);
-    Task<Tournament?> GetTournamentByIdAsync(int tournamentId);
-    Task<Race?> GetFinalRaceInTournamentAsync(int tournamentId);
+    Task<Tournament?> GetTournamentByIdAsync(long tournamentId);
+    Task<Race?> GetFinalRaceInTournamentAsync(long tournamentId);
 
     // Prediction queries
-    Task<Prediction?> GetPredictionAsync(int raceId, int userId);
-    Task<IEnumerable<Prediction>> GetPredictionsByRaceIdAsync(int raceId);
+    Task<Prediction?> GetPredictionAsync(long raceId, int userId);
+    Task<IEnumerable<Prediction>> GetPredictionsByRaceIdAsync(long raceId);
     Task AddPredictionAsync(Prediction prediction);
 
     // Rankings queries
