+++ backend/src/HorseRacing.Domain/Entities/Financials/Prize.cs
@@ -1,9 +1,11 @@
+using HorseRacing.Domain.Entities.Tournaments;
+
 namespace HorseRacing.Domain.Entities;
 
 public class Prize
 {
     public int Id { get; set; }
-    public int TournamentId { get; set; }
+    public long TournamentId { get; set; }
     public Tournament? Tournament { get; set; }
     public int Rank { get; set; } // 1, 2, 3
     public decimal Amount { get; set; }
